requirements.txt
# requirements.txt - Updated versions for 2024/2025
flask==3.0.3
flask-cors==5.0.0
falkordb==1.0.9
qdrant-client==1.11.3
python-dotenv==1.0.1
python-dateutil==2.9.0
openai==1.55.3
spacy==3.8.7
requests==2.31.0
fastembed==0.4.2
onnxruntime<1.20  # Pin to avoid issues with fastembed 0.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz
httpx>=0.27.0
